Author: Roger Heflin (rogerheflin@gmail.com)
Changes to reduce the size of the timestamp structure.

The original structure is 7 - longs, the new structure is 1 short int, and
6 unsigned chars, reducing the structure size by 20 bytes, there are 6
timestamps per recording so this reduces the size of *EACH* recording's
data structure by 120 bytes (average structure size on my testing went from 1000
bytes down to 880 bytes), and assuming we have 3 MB useable (2 copies at
1.5MB each during update), this increases the number of recordngs that can be
used from 1500 to 1700.

+++ libs/libcmyth/timestamp.c
@@ -303,12 +303,12 @@ cmyth_timestamp_to_string(char *str, cmyth_timestamp_t ts)
 	}
 	sprintf(str,
 		"%4.4ld-%2.2ld-%2.2ldT%2.2ld:%2.2ld:%2.2ld",
-		ts->timestamp_year,
-		ts->timestamp_month,
-		ts->timestamp_day,
-		ts->timestamp_hour,
-		ts->timestamp_minute,
-		ts->timestamp_second);
+		(unsigned long)ts->timestamp_year,
+		(unsigned long)ts->timestamp_month,
+		(unsigned long)ts->timestamp_day,
+		(unsigned long)ts->timestamp_hour,
+		(unsigned long)ts->timestamp_minute,
+		(unsigned long)ts->timestamp_second);
 	return 0;
 }
 
@@ -345,9 +345,9 @@ cmyth_timestamp_to_isostring(char *str, cmyth_timestamp_t ts)
 	}
 	sprintf(str,
 		"%4.4ld-%2.2ld-%2.2ld",
-		ts->timestamp_year,
-		ts->timestamp_month,
-		ts->timestamp_day);
+		(unsigned long)ts->timestamp_year,
+		(unsigned long)ts->timestamp_month,
+		(unsigned long)ts->timestamp_day);
 	return 0;
 }
 
@@ -379,24 +379,24 @@ cmyth_timestamp_to_display_string(char *str, cmyth_timestamp_t ts,
 
 		sprintf(str,
 			"%4.4ld-%2.2ld-%2.2ldT%2.2ld:%2.2ld:%2.2ld %s",
-			ts->timestamp_year,
-			ts->timestamp_month,
-			ts->timestamp_day,
+			(unsigned long)ts->timestamp_year,
+			(unsigned long)ts->timestamp_month,
+			(unsigned long)ts->timestamp_day,
 			hour,
-			ts->timestamp_minute,
-			ts->timestamp_second,
+			(unsigned long)ts->timestamp_minute,
+			(unsigned long)ts->timestamp_second,
 			pm ? "PM" : "AM");
 	}
 	else
 	{
 		sprintf(str,
 			"%4.4ld-%2.2ld-%2.2ldT%2.2ld:%2.2ld:%2.2ld",
-			ts->timestamp_year,
-			ts->timestamp_month,
-			ts->timestamp_day,
-			ts->timestamp_hour,
-			ts->timestamp_minute,
-			ts->timestamp_second);
+			(unsigned long)ts->timestamp_year,
+			(unsigned long)ts->timestamp_month,
+			(unsigned long)ts->timestamp_day,
+			(unsigned long)ts->timestamp_hour,
+			(unsigned long)ts->timestamp_minute,
+			(unsigned long)ts->timestamp_second);
 	}
 	return 0;
 }
@@ -447,12 +447,12 @@ cmyth_datetime_to_string(char *str, cmyth_timestamp_t ts)
 	t_datetime = mktime(&tm_datetime);
 	sprintf(str,
 		"%4.4ld-%2.2ld-%2.2ldT%2.2ld:%2.2ld:%2.2ld",
-		ts->timestamp_year,
-		ts->timestamp_month,
-		ts->timestamp_day,
-		ts->timestamp_hour,
-		ts->timestamp_minute,
-		ts->timestamp_second);
+		(unsigned long)ts->timestamp_year,
+		(unsigned long)ts->timestamp_month,
+		(unsigned long)ts->timestamp_day,
+		(unsigned long)ts->timestamp_hour,
+		(unsigned long)ts->timestamp_minute,
+		(unsigned long)ts->timestamp_second);
 	cmyth_dbg(CMYTH_DBG_ERROR, "original timestamp string: %s \n",str);
 	sprintf(str,"%lu",(unsigned long) t_datetime);
 	cmyth_dbg(CMYTH_DBG_ERROR, "time in seconds: %s \n",str);
